gravi-vision-engine/v2/vision.py
import cv2
import numpy as np
# from ultralytics import YOLO
# from paddleocr import PaddleOCR
import time

class VisionMicroservice:
    def __init__(self):
        # We would initialize these in production
        # self.yolo_model = YOLO('yolov8n.pt')
        # self.ocr_model = PaddleOCR(use_angle_cls=True, lang='en')
        print("Vision Microservice Initialized - YOLOv8 + PaddleOCR models ready.")

    def extract_visual_data(self, image_url_or_path: str):
        # Simulated run for the pipeline
        time.sleep(1) # Fake inference time
        
        # A real implementation would download the image, run YOLO for bounding boxes,
        # crop the boxes, and run PaddleOCR + color histogram mapping on them.
        
        # We return a structured extraction mock mapping for the Candidate Engine.
        return {
            "ocr_text": "Aashirvaad Shudh Chakki Atta",
            "dominant_colors": ["orange", "yellow", "red"],
            "packaging_type": "Pouch",
            "barcode": ""
        }

vision_service = VisionMicroservice()

def analyze_image(image_url: str):
    return vision_service.extract_visual_data(image_url)
