gravi-vision-engine/v2/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
thefuzz>=0.22.1
requests>=2.32.3
pydantic>=2.7.4
python-multipart>=0.0.9
